# specs/openapi/adscenter.yaml
openapi: 3.0.3
info:
  title: Adscenter Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /
paths:
  /api/v1/adscenter/accounts:
    get:
      operationId: listAccounts
      summary: List accessible ad accounts
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
    post:
      operationId: addAccount
      summary: Add a Google Ads account reference for current user (metadata only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId: { type: string }
                accountName: { type: string }
              required: [accountId]
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/accounts/{id}:
    delete:
      operationId: deleteAccount
      summary: Delete an account reference
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/accounts/stream:
    get:
      operationId: streamAccounts
      summary: Stream accounts with real-time updates (SSE)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            text/event-stream:
              schema:
                type: string
        '401': { description: Unauthorized }
  /api/v1/adscenter/accounts/sync-all:
    post:
      operationId: syncAllAccounts
      summary: Sync all accounts from Google Ads API
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  taskId: { type: string }
                  status: { type: string }
        '401': { description: Unauthorized }
        '501': { description: Not Implemented }
...
  /api/v1/adscenter/execution-report:
    get:
      operationId: getExecutionReport
      summary: Get execution report for campaigns
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          required: false
          schema: { type: integer, default: 7 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  executions:
                    type: array
                    items: { type: object }
        '401': { description: Unauthorized }
  /api/v1/adscenter/budget/transfer:
    post:
      operationId: transferBudget
      summary: Transfer budget between accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_account_id: { type: string }
                to_account_id: { type: string }
                amount: { type: number }
              required: [from_account_id, to_account_id, amount]
      responses:
        '200': { description: Transfer successful }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/configurations:
    get:
      operationId: listConfigurations
      summary: List user configurations for Adscenter
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Configuration' }
        '401': { description: Unauthorized }
    post:
      operationId: createConfiguration
      summary: Create a configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                payload:
                  type: object
                  additionalProperties: true
              required: [name]
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/configurations/{id}:
    delete:
      operationId: deleteConfiguration
      summary: Delete a configuration by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/executions:
    post:
      operationId: startExecution
      summary: Start an execution with given configuration id
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                configurationId: { type: string }
              required: [configurationId]
      responses:
        '202': { description: Accepted }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/preflight:
    post:
      operationId: runPreflight
      summary: Run pre-flight diagnostics
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId:
                  type: string
                  description: Google Ads customer id (10 digits)
                validateOnly:
                  type: boolean
                  description: If true, skip live checks and only validate environment/config
                landingUrl:
                  type: string
                  description: Optional landing page URL to check reachability
              required: []
      responses:
        '200':
          description: Diagnostics result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResult'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/bulk-actions:
    get:
      operationId: listBulkActions
      summary: List recent bulk action operations for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BulkActionOperation'
        '401': { description: Unauthorized }
    post:
      operationId: submitBulkActions
      summary: Submit bulk actions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/BulkActionPlan'
                - $ref: '#/components/schemas/OpportunityComboPlan'
      responses:
        '200': { description: OK (validateOnly) }
        '202': { description: Accepted (enqueued) }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/bulk-actions/{id}:
    get:
      operationId: getBulkAction
      summary: Get bulk action operation status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionOperation'
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/{id}/plan:
    get:
      operationId: getBulkActionPlan
      summary: Get the submitted plan body of a bulk action
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: object
                    additionalProperties: true
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/validate:
    post:
      operationId: validateBulkActions
      summary: Validate a bulk action plan without enqueuing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkActionPlan'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionValidationResult'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/bulk-actions/{id}/audits:
    get:
      operationId: getBulkActionAudits
      summary: List audit snapshots for an operation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BulkActionAuditItem'
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/{id}/rollback:
    post:
      operationId: rollbackBulkAction
      summary: Mark operation as rolled back and append a rollback snapshot (stub)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rolled back }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/{id}/rollback-plan:
    post:
      operationId: getRollbackPlan
      summary: Generate a rollback plan (inverse actions) without executing
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/BulkActionPlan'
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/{id}/rollback-execute:
    post:
      operationId: rollbackExecute
      summary: Execute rollback (stub; writes audits only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  executed: { type: integer }
                  errors: { type: integer }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/bulk-actions/{id}/report:
    get:
      operationId: getRollbackReport
      summary: Get rollback execution report (audit snapshots)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: kind
          schema: { type: string, enum: [rollback, rollback_exec] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BulkActionAuditItem'
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/oauth/url:
    get:
      operationId: getOAuthUrl
      summary: Get OAuth URL
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /api/v1/adscenter/oauth/callback:
    get:
      operationId: oauthCallback
      summary: OAuth callback
      security: []
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/oauth/revoke:
    post:
      operationId: oauthRevoke
      summary: Revoke stored Google Ads refresh token for current user
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /api/v1/adscenter/mcc/link:
    post:
      operationId: mccLink
      summary: Send manager link invitation (stub/live)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId: { type: string }
              required: [customerId]
      responses:
        '200': { description: OK }
        '202': { description: Accepted (idempotent queued) }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/mcc/status:
    get:
      operationId: mccStatus
      summary: Get manager link status
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: customerId
          schema: { type: string }
          required: true
      responses:
        '200': { description: OK }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/v1/adscenter/mcc/unlink:
    post:
      operationId: mccUnlink
      summary: Request unlink from manager account (stub/live)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId: { type: string }
              required: [customerId]
      responses:
        '200': { description: OK }
        '202': { description: Accepted (idempotent queued) }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/mcc/refresh:
    post:
      operationId: mccRefresh
      summary: Refresh statuses for all pending/invited MCC links (best-effort)
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                shard: { type: integer, minimum: 0 }
                totalShards: { type: integer, minimum: 1 }
              required: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  checked: { type: integer }
                  updated: { type: integer }
                  live: { type: boolean }
        '401': { description: Unauthorized }
  /api/v1/adscenter/mcc/links:
    get:
      operationId: listMccLinks
      summary: List MCC links for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [invited, pending, active, inactive] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MccLink'
        '401': { description: Unauthorized }
  /api/v1/adscenter/connections:
    get:
      operationId: listAdsConnections
      summary: List stored Google Ads connections (sanitized)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdsConnection'
        '401': { description: Unauthorized }
  /api/v1/adscenter/audits:
    get:
      operationId: listAuditEvents
      summary: List audit events
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: kind
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
        - in: query
          name: since
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEventItem'
        '401': { description: Unauthorized }
  /api/v1/adscenter/keywords/expand:
    post:
      operationId: expandKeywords
      summary: Expand keywords using seed domain/keywords
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seedDomain: { type: string }
                seedKeywords:
                  type: array
                  items: { type: string }
                validateOnly:
                  type: boolean
                  description: If true, returns stub suggestions without live API calls
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/KeywordIdea'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/diagnose:
    post:
      operationId: diagnose
      summary: Diagnose ad account issues based on provided metrics and landing URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId: { type: string }
                landingUrl: { type: string }
                metrics:
                  type: object
                  additionalProperties: true
              required: [accountId]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnoseResult'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/diagnose/plan:
    post:
      operationId: diagnosePlan
      summary: Generate a validate-only bulk action plan from metrics
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId: { type: string }
                metrics:
                  type: object
                  additionalProperties: true
                suggestedActions:
                  type: array
                  items: { $ref: '#/components/schemas/SuggestedAction' }
              required: [accountId, metrics]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionPlan'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
...
  /api/v1/adscenter/diagnose/execute:
    post:
      operationId: diagnoseExecute
      summary: Generate a plan from metrics and enqueue as a bulk operation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metrics:
                  type: object
                  additionalProperties: true
              required: [metrics]
      responses:
        '202': { description: Accepted }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /api/v1/adscenter/limits/me:
    get:
      operationId: getLimitsMe
      summary: Get current plan limits and daily quota usage for the user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan: { type: string }
                  limits:
                    type: object
                    properties:
                      preflight:
                        type: object
                        properties:
                          rpm: { type: integer }
                          concurrency: { type: integer }
                      mutate:
                        type: object
                        properties:
                          rpm: { type: integer }
                          concurrency: { type: integer }
                  quota:
                    type: object
                    properties:
                      daily: { type: integer }
                      usedToday: { type: integer }
        '401': { description: Unauthorized }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Configuration:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        payload:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PreflightResult:
      type: object
      properties:
        summary: { type: string, enum: [ok, warn, error] }
        checks:
          type: array
          items: { $ref: '#/components/schemas/PreflightCheck' }
      required: [summary, checks]
    BulkActionOperation:
      type: object
      properties:
        operationId: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        summary:
          type: object
          properties:
            actions: { type: integer }
            estimatedAffected: { type: integer }
      required: [operationId, status]
    BulkActionPlan:
      type: object
      properties:
        validateOnly:
          type: boolean
        actions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [ADJUST_CPC, ADJUST_BUDGET, ADD_NEGATIVE_KEYWORDS, REMOVE_NEGATIVE_KEYWORDS, PAUSE_KEYWORDS, ENABLE_KEYWORDS, SET_AD_SCHEDULES, SET_TARGET_CPA, SET_TARGET_ROAS]
              filter:
                type: object
                additionalProperties: true
              params:
                oneOf:
                  - $ref: '#/components/schemas/AdjustCpcParams'
                  - $ref: '#/components/schemas/AdjustBudgetParams'
                  - $ref: '#/components/schemas/NegativeKeywordsParams'
                  - $ref: '#/components/schemas/KeywordStatusParams'
                  - $ref: '#/components/schemas/AdSchedulesParams'
                  - $ref: '#/components/schemas/BiddingParams'
          minItems: 1
      required: [actions]
...
    OpportunityComboPlan:
      type: object
      description: Minimal combo plan built from an Opportunity (keywords/domains) without enumerating low-level actions.
      properties:
        validateOnly: { type: boolean }
        seedDomain: { type: string }
        country: { type: string }
        plan:
          type: object
          properties:
            keywords:
              type: array
              items:
                type: object
                properties:
                  keyword: { type: string }
                  score: { type: number }
                  reason: { type: string }
            domains:
              type: array
              items:
                type: object
                properties:
                  domain: { type: string }
                  score: { type: number }
...